Automatically build and push Docker container (#61)
Closes #52.

This PR adds a workflow to automatically build and push the PyGEM image to the GitHub Container Registry (ghcr.io), that will then be visible at https://github.com/orgs/PyGEM-Community/packages. It builds it from the Dockerfile that I add as well.

It's pushed to ghcr only when the workflow is triggered on the master or dev branches: that is, if it's triggered by schedule or when we merge relevant commits to the branches. But it's always built in PRs so we can check that it builds successfully.

Currently the image is quite large (9.1 GB) but can be reduced to about 2.5GB, as described in #60.