Run tests as part of CI (#59)
Closes #53.

This PR adds a workflow that runs the test suite every Saturday and every time commits are pushed to open PRs or to the master and dev branches. In the interest of green computing, it is only triggered if these commits change at least one of these files:
- any `.py` file
- the workflow itself, i.e., `.github/workflows/test_suite.yml`
- `pyproject.toml` so that we can test the installation

Deleted `test_oggm_compat.py` since all tests fail anyway.