Use Ruff linter and formatter (#62)
Closes #54.

This PR introduces the [Ruff](https://github.com/astral-sh/ruff) linter and formatter to check for errors and stylistic issues, including the pre-commit hook. Relevant changes are done in `pyproject.toml` and `test_suite.yml`. The rest of changes are due to applying the fixes. 

A few things to note:
- I updated contribution guidelines
- I added the ruff linting and formatting checks to the `test_suite.yml` workflow
- The maximum line length is set to 88 by default, which we can change if you'd prefer
- At the moment I ignored quite a few errors that couldn't be fixed automatically (see the `ignore` list in `pyproject.toml`) - I'll open a separate issue to address these. **Edit: opened #68** 
- I added `sample_data/` to `.gitignore`