Remove assertion that bias adjustment precipitation factor be between 0.5 and 2
Closes #118. 
This PR returns the GCM precipitation bias adjustment factors to `run_simulation.py` and removed previous assertions that the bias-adjusted GCM precipitation divided by the reference climate precipitation not differ by a factor >2 for the same time period.

A warning is now raised in `run_simulation.py` if the factor differs by >2x, but the simulation will still run.

Also cleaned up relict variable naming that was in place to handle `ref_spinupyears` and `gcm_spinupyears`.