Bug fix in exporting extra variables, added MCMC stats (#123)
Closes #105.
This PR adds a function to calculate some basic statistics on the MCMC model parameters in `run_calibration`. These stats are now added to the output calibration json file under the '_summary_stats_' key. This PR also patches a minor bug in the output of `run_simulation` when the `export_extra_vars` option is selected. Previously, if `export_extra_vars` was not set as `True` in PyGEM's configuration file, an error would be thrown. Now, `pygem.output` takes `export_extra_vars` from the command-line arguments - with defaults taken from the configuration file.